pyproject.toml
[tool.poetry]
name = "codd-testovoe"
version = "0.1.0"
description = ""
authors = ["Tagir_Mirzametov <tm@uktlab.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.95.1"
uvicorn = "^0.21.1"
asyncpg = "^0.27.0"
aiosql = "^8.0"
pydantic = {extras = ["dotenv"], version = "^1.10.7"}
alembic = "^1.10.3"
psycopg2-binary = "^2.9.6"
beautifulsoup4 = "^4.12.2"
lxml = "^4.9.2"
fastapi-utils = "^0.2.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
